% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_node_color.R
\name{apply_node_color}
\alias{apply_node_color}
\title{Apply node colors based on the node's type}
\usage{
apply_node_color(p)
}
\arguments{
\item{p}{process map generated \code{by create_pmap()}}
}
\description{
Different type of node should be differenciate by different colors, this function applies the material design palette for the node's color.
}
