% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph.R
\name{create_event_graph}
\alias{create_event_graph}
\title{Create the event graph by given nodes and edges.}
\usage{
create_event_graph(nodes, edges)
}
\arguments{
\item{nodes}{Event list, which should be a `data.frame` containing following columns:
* `name`: Nodes name, will be used as label. (`character`)
* `is_target`: Whether it's final stage. (`logical`)
* `percentage`: The percentage of customer affected by the given event. (`numeric`)}

\item{edges}{Event transform list, which should be a `data.frame` containing following columns:
* `from`: the begining event of the edge. (`character`)
* `to`: the ending event of the edge (`character`)
* `value`: How many of customer affected by the given event. (`numeric`)}
}
