% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pmap.R
\name{create_pmap}
\alias{create_pmap}
\title{Create the event graph by given nodes and edges.}
\usage{
create_pmap(nodes, edges, render = T)
}
\arguments{
\item{nodes}{Event list}

\item{edges}{Event transform list}

\item{render}{Whether should call \code{render_graph()} on the final result. Set to \code{FALSE} if you want further manipulate the result by \code{DiagrammeR} functions.}
}
\description{
\code{nodes} should be a \code{data.frame} containing following columns:
\itemize{
\item \code{name}: Nodes name, will be used as label. (\code{character})
\item \code{is_target}: Whether it's final stage. (\code{logical})
}

\code{edges} should be a \code{data.frame} containing following columns:
\itemize{
\item \code{from}: the begining event of the edge. (\code{character})
\item \code{to}: the ending event of the edge (\code{character})
\item \code{value}: How many of customer affected by the given event. (\code{numeric})
}
}
