% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_edges.R
\name{generate_edges}
\alias{generate_edges}
\title{Generate edges from event logs}
\usage{
generate_edges(eventlog)
}
\arguments{
\item{eventlog}{Event logs}
}
\value{
a \code{data.frame} of edges
}
\description{
\code{eventlog} should be a \code{data.frame} or \code{data.table}, which contains, at least, following columns:
\itemize{
\item \code{timestamp}: timestamp column which indicates when event happened. (\code{POSIXct})
\item \code{customer_id}: cutomer identifier. (\code{character})
\item \code{event_name}: event name. (\code{character})
\item \code{is_target}: whether it's the final stage. (\code{logical})
}
}
