% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nodes.R
\name{generate_nodes}
\alias{generate_nodes}
\title{Generate nodes from event logs}
\usage{
generate_nodes(eventlog)
}
\arguments{
\item{eventlog}{Event logs}
}
\value{
a \code{data.frame} of nodes
}
\description{
\code{eventlog} should be a \code{data.frame}, which contains, at least, following columns:
\itemize{
\item \code{event_name}: event name. (\code{character})
\item \code{is_target}: whether it's the final stage. (\code{logical})
}

\code{generate_nodes()} will generate the node list from the given \code{eventlog} for the graph purpose.
}
