% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_eventlog.R
\name{generate_random_eventlog}
\alias{generate_random_eventlog}
\title{Generate Random Eventlog}
\usage{
generate_random_eventlog(
     size_of_eventlog = 1000, 
     number_of_customers = 20, 
     event_catalogs = c("normal", "target"), 
     event_catalogs_size = c(8, 2))
}
\arguments{
\item{size_of_eventlog}{The size of generated event log}

\item{number_of_customers}{How many customers in the simulation}

\item{event_catalogs}{A data frame contains the event catelog}

\item{event_catalogs_size}{How many event types in each event catalog}
}
\description{
This function provides the ability to randomly generate the eventlog data frame based on given parameters.
}
\details{
Example\preformatted{> eventlog <- generate_random_eventlog(500, 20)
> str(eventlog)
'data.frame':	500 obs. of  4 variables:
 $ timestamp  : POSIXct, format: "2017-01-01 03:54:47" "2017-01-01 20:31:02" ...
 $ customer_id: chr  "Customer 20" "Customer 20" "Customer 1" "Customer 3" ...
 $ event_name : chr  "Event 10 (target)" "Event 6 (normal)" "Event 8 (normal)" "Event 9 (target)" ...
 $ event_type : chr  "target" "normal" "normal" "target" ...
}
}
