% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attrs_desc.R
\name{get_attrs_desc}
\alias{get_attrs_desc}
\title{Get an attribute key-pair string from an object}
\usage{
get_attrs_desc(object)
}
\arguments{
\item{object}{Given object, can be \code{list}, \code{matrix}, or \code{data.frame}}
}
\description{
Get an attribute list string from an object
}
