% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph.R
\name{get_edges_from_event_logs}
\alias{get_edges_from_event_logs}
\title{Produce edges from event logs}
\usage{
get_edges_from_event_logs(event_logs)
}
\arguments{
\item{event_logs}{Event logs}
}
\description{
`event_logs` should be a `data.frame` or `data.table`, which contains, at least, following columns:

 * `timestamp`: timestamp column which indicates when event happened. (`POSIXct`)
 * `customer_id`: cutomer identifier. (`character`)
 * `event_name`: event name. (`character`)
 * `is_target`: whether it's the final stage. (`logical`)
}
