% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_edges.R
\name{prune_edges}
\alias{prune_edges}
\title{Prune edges based on given percentage}
\usage{
prune_edges(p, percentage = 0.2)
}
\arguments{
\item{p}{process map object created by \code{create_pmap()} function}

\item{percentage}{how many percentage of the edges should be pruned.}
}
\description{
Prune edges based on given percentage
}
