% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pmap.R
\name{render_pmap}
\alias{render_pmap}
\title{Render the process map}
\usage{
render_pmap(p, title = NULL)
}
\arguments{
\item{p}{the process map object created by \code{create_pmap()} function}

\item{title}{The title of rendered graph}
}
\description{
Basically, this function just called \code{DiagrammeR::render_graph()}
}
