% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pmap_shiny.R
\name{render_pmap_shiny}
\alias{render_pmap_shiny}
\title{Render process map as a web app}
\usage{
render_pmap_shiny(p, title = "Process Map", options = NULL)
}
\arguments{
\item{p}{The process map created by \code{create_pmap()}}

\item{title}{The title you want to display on the web page}

\item{options}{The Shiny server options, such as binding address or listening port.}
}
\description{
Show the given process map in a Shiny server web application, with ability to adjust the nodes and edges precision in the real time.
}
\examples{
# We can use the real life event log from 'eventdataR' package to do the shiny app demo
library(eventdataR)
library(dplyr)
eventlog <- eventdataR::sepsis \%>\%
  rename(
    timestamp = Complete_Timestamp,
    customer_id = Case_ID,
    event_name = Activity
  ) \%>\%
  mutate(
    event_type = event_name
  ) \%>\%
  select(timestamp, customer_id, event_name, event_type) \%>\%
  filter(!is.na(customer_id))
app <- eventlog \%>\% create_pmap() \%>\% render_pmap_shiny()
# To run the app, just print(app)
}
