% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph.R
\name{create_event_graph}
\alias{create_event_graph}
\title{Create the event graph by given nodes and edges.}
\usage{
create_event_graph(nodes, edges)
}
\arguments{
\item{nodes}{Event list}

\item{edges}{Event transform list}
}
\description{
\code{nodes} should be a \code{data.frame} containing following columns:
\itemize{
\item \code{name}: Nodes name, will be used as label. (\code{character})
\item \code{is_target}: Whether it's final stage. (\code{logical})
\item \code{percentage}: The percentage of customer affected by the given event. (\code{numeric})
}

\code{edges} should be a \code{data.frame} containing following columns:
\itemize{
\item \code{from}: the begining event of the edge. (\code{character})
\item \code{to}: the ending event of the edge (\code{character})
\item \code{value}: How many of customer affected by the given event. (\code{numeric})
}
}
