% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pmap.R
\name{create_pmap}
\alias{create_pmap}
\title{Create the event graph by given nodes and edges.}
\usage{
create_pmap(nodes, edges, target_types = NULL)
}
\arguments{
\item{nodes}{Event list}

\item{edges}{Event transform list}

\item{target_types}{A vector contains the target event types}
}
\description{
\code{nodes} should be a \code{data.frame} containing following columns:
\itemize{
\item \code{name}: Event name, will be used as label. (\code{character})
\item \code{type}: The event type (\code{character})
}

\code{edges} should be a \code{data.frame} containing following columns:
\itemize{
\item \code{from}: the begining event of the edge. (\code{character})
\item \code{to}: the ending event of the edge (\code{character})
\item \code{amount}: How many of customer affected by the given event. (\code{numeric})
}
}
\seealso{
\link{create_pmap_from_eventlog}
}
