% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pmap_from_eventlog.R
\name{create_pmap_from_eventlog}
\alias{create_pmap_from_eventlog}
\title{Create the process map from event log directly}
\usage{
create_pmap_from_eventlog(eventlog)
}
\arguments{
\item{eventlog}{Event log}
}
\description{
Create the process map by analysing given \code{eventlog} and extract the nodes by \code{generate_nodes()} and edges by \code{generate_edges()}.
}
