% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_edges.R
\name{generate_edges}
\alias{generate_edges}
\title{Generate edges from event logs}
\usage{
generate_edges(eventlog, distinct_customer = F, target_types = NULL)
}
\arguments{
\item{eventlog}{Event logs}

\item{distinct_customer}{Whether should only count unique customer}

\item{target_types}{A vector contains the target event types. By default, it's \code{NULL}, which means every paths count. If it's contains the target event type, then only paths reaches the target event count.}
}
\value{
a \code{data.frame} of edges with \code{from}, \code{to} and \code{amount} columns.
}
\description{
\code{eventlog} should be a \code{data.frame} or \code{data.table}, which contains, at least, following columns:
\itemize{
\item \code{timestamp}: timestamp column which indicates when event happened. (\code{POSIXct})
\item \code{customer_id}: customer identifier. (\code{character})
\item \code{event_name}: event name. (\code{character})
\item \code{event_type}: event type. (\code{character})
}
}
