% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nodes.R
\name{generate_nodes}
\alias{generate_nodes}
\title{Generate nodes from event logs}
\usage{
generate_nodes(eventlog, distinct_customer = F)
}
\arguments{
\item{eventlog}{Event logs}

\item{distinct_customer}{Whether should only count unique customer}
}
\value{
a nodes \code{data.frame} which represents a event list, it contains \code{name}, \code{type} and \code{amount} columns.
}
\description{
\code{eventlog} should be a \code{data.frame}, which contains, at least, following columns:
\itemize{
\item \code{event_name}: event name. (\code{character})
\item \code{event_type}: event type. (\code{character})
\item \code{amount}: how many time this event happened in the \code{eventlog}
}

\code{generate_nodes()} will generate the node list from the given \code{eventlog} for the graph purpose.
}
