% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attrs_desc.R
\name{get_attrs_desc}
\alias{get_attrs_desc}
\title{Get an attribute key-pair string from an object}
\usage{
get_attrs_desc(object)
}
\arguments{
\item{object}{Given object, can be \code{list}, \code{matrix}, or \code{data.frame}}
}
\description{
Get an attribute list string from an object
}
\examples{
print(df)
#   id   name is_manager
# 1  1   Jane      FALSE
# 2  2   John      FALSE
# 3  3   Eric      FALSE
# 4  4 Selena       TRUE
get_attrs_desc(df)
# [1] "id: 1\\nname: Jane\\nis_manager: FALSE"
# [2] "id: 2\\nname: John\\nis_manager: FALSE"
# [3] "id: 3\\nname: Eric\\nis_manager: FALSE"
# [4] "id: 4\\nname: Selena\\nis_manager: TRUE"
}
