% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph.R
\name{get_edges_from_event_logs}
\alias{get_edges_from_event_logs}
\title{Produce edges from event logs}
\usage{
get_edges_from_event_logs(event_logs)
}
\arguments{
\item{event_logs}{Event logs}
}
\value{
a \code{data.frame} of edges
}
\description{
\code{event_logs} should be a \code{data.frame} or \code{data.table}, which contains, at least, following columns:
\itemize{
\item \code{timestamp}: timestamp column which indicates when event happened. (\code{POSIXct})
\item \code{customer_id}: cutomer identifier. (\code{character})
\item \code{event_name}: event name. (\code{character})
\item \code{is_target}: whether it's the final stage. (\code{logical})
}
}
