% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graph.R
\name{get_nodes_from_event_logs}
\alias{get_nodes_from_event_logs}
\title{Produce nodes from event logs}
\usage{
get_nodes_from_event_logs(event_logs)
}
\arguments{
\item{event_logs}{Event logs}
}
\value{
a \code{data.frame} of nodes
}
\description{
\code{event_logs} should be a \code{data.frame} or \code{data.table}, which contains, at least, following columns:
\itemize{
\item \code{event_name}: event name. (\code{character})
\item \code{is_target}: whether it's the final stage. (\code{logical})
}

\code{get_nodes_from_event_logs()} will generate the node list from the given \code{event_logs} for the graph purpose.
However, it is very basic, so, in most cases, you might want to provide your own node list to the
\code{create_event_graph()} function.
}
