% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pmap.R
\name{render_pmap}
\alias{render_pmap}
\title{Render the process map}
\usage{
render_pmap(p, title = NULL)
}
\arguments{
\item{p}{the process map object created by \code{create_pmap_graph()} function}

\item{title}{The title of rendered graph}
}
\description{
Basically, this function just called \code{DiagrammeR::render_graph()}
}
\examples{
library(dplyr)
p <- generate_eventlog() \%>\% create_pmap(target_types = c("target"))
render_pmap
}
\seealso{
\link{create_pmap}
}
